+++ main.py
@@ -14,118 +14,6 @@ from tinydb import TinyDB
 from packages.helper  import Helper
 import os
 
-class NeuerKursDialog(QtWidgets.QDialog):
-    def __init__(self):
-        super().__init__()
-        self.setWindowTitle("Neuer Kurs")
-
-        layout = QtWidgets.QVBoxLayout()
-        self.setFixedSize(600, 800)  # Set fixed size for the dialog window
-
-        self.setWhatsThis("Dieser Dialog ermöglicht es dir, ein neuen kurs einzutragen. "
-                          "Wähle den Kurs, das kurs Rating, das Slope Rating, die Anzahl der Löcher und gib die Par Schlagzahlen ein.")
-        self.kurs_name_eingabe = QtWidgets.QLineEdit()
-        layout.addWidget(QtWidgets.QLabel("Kursname:"))
-        layout.addWidget(self.kurs_name_eingabe)
-
-        self.kurs_rating_eingabe = QtWidgets.QLineEdit()
-        layout.addWidget(QtWidgets.QLabel("Kurs Rating:"))
-        layout.addWidget(self.kurs_rating_eingabe)
-
-        self.slope_rating_eingabe = QtWidgets.QLineEdit()
-        layout.addWidget(QtWidgets.QLabel("Slope Rating:"))
-        layout.addWidget(self.slope_rating_eingabe)
-
-        self.neun_loch_radio = QtWidgets.QRadioButton("9 Löcher")
-        self.achtzehn_loch_radio = QtWidgets.QRadioButton("18 Löcher")
-        self.neun_loch_radio.setChecked(True)
-        layout.addWidget(QtWidgets.QLabel("Anzahl Löcher:"))
-        layout.addWidget(self.neun_loch_radio)
-        layout.addWidget(self.achtzehn_loch_radio)
-
-        schlagzahl_layout = QtWidgets.QGridLayout()
-
-        self.par_eingabe = {}
-        self.par_labels = {}
-        for i in range(1, 19):
-            eingabe = QtWidgets.QLineEdit()
-            eingabe.setValidator(QtGui.QIntValidator())
-            self.par_eingabe[i] = eingabe
-            label_text = f"Par Loch {i}:"
-            label = QtWidgets.QLabel(label_text)
-            self.par_labels[i] = label
-
-            row = (i - 1) % 9
-            col = (i - 1) // 9
-            schlagzahl_layout.addWidget(label, row, col * 2)
-            schlagzahl_layout.addWidget(eingabe, row, col * 2 + 1)
-
-            if i > 9:
-                eingabe.hide()
-                label.hide()
-
-        layout.addLayout(schlagzahl_layout)
-
-        stroke_index_layout = QtWidgets.QGridLayout()
-
-        self.stroke_index_eingabe = {}
-        self.stroke_index_labels = {}
-        for i in range(1, 19):
-            eingabe = QtWidgets.QLineEdit()
-            eingabe.setValidator(QtGui.QIntValidator())
-            self.stroke_index_eingabe[i] = eingabe
-            label_text = f"Stroke Index Loch {i}:"
-            label = QtWidgets.QLabel(label_text)
-            self.stroke_index_labels[i] = label
-
-            row = (i - 1) % 9  # Reset row for stroke index inputs
-            col = (i - 1) // 9
-            stroke_index_layout.addWidget(label, row, col * 2)
-            stroke_index_layout.addWidget(eingabe, row, col * 2 + 1)
-
-            if i > 9:
-                eingabe.hide()
-                label.hide()
-
-        layout.addLayout(stroke_index_layout)
-
-        button_box = QtWidgets.QDialogButtonBox(QtWidgets.QDialogButtonBox.Ok | QtWidgets.QDialogButtonBox.Cancel)
-        button_box.accepted.connect(self.accept)
-        button_box.rejected.connect(self.reject)
-        layout.addWidget(button_box)
-
-        self.setLayout(layout)
-
-        self.neun_loch_radio.toggled.connect(self.update_par_eingabe)
-        self.achtzehn_loch_radio.toggled.connect(self.update_par_eingabe)
-
-    def update_par_eingabe(self):
-        locher = 9 if self.neun_loch_radio.isChecked() else 18
-        for i in range(1, 19):
-            self.par_eingabe[i].setVisible(i <= locher)
-            self.par_labels[i].setVisible(i <= locher)
-            self.stroke_index_eingabe[i].setVisible(i <= locher)
-            self.stroke_index_labels[i].setVisible(i <= locher)
-
-    def get_kurs_daten(self):
-        kurs_name = self.kurs_name_eingabe.text()
-        kurs_rating = self.kurs_rating_eingabe.text()
-        slope_rating = self.slope_rating_eingabe.text()
-        locher = 9 if self.neun_loch_radio.isChecked() else 18
-        pars = []
-        stroke_indices = []
-        for i in range(1, locher + 1):
-            try:
-                pars.append(int(self.par_eingabe[i].text()))
-            except ValueError:
-                pars.append(0)
-            try:
-                stroke_indices.append(int(self.stroke_index_eingabe[i].text()))
-            except ValueError:
-                stroke_indices.append(0)
-
-        help.addCourse(kurs_name, int(kurs_rating), int(slope_rating), pars, stroke_indices)
-
 class KursLoeschenDialog(QtWidgets.QDialog):
     def __init__(self, kurse):
         super().__init__()
@@ -276,10 +164,8 @@ class HandicapUI(QtWidgets.QWidget):
         
 
         button_header = QtWidgets.QHBoxLayout()
-        self.neuer_kurs_button = QtWidgets.QPushButton("Neuer Kurs")
         self.kurs_löschen_button = QtWidgets.QPushButton("Kurs Löschen")
         self.export_scorecard_button = QtWidgets.QPushButton("Scorecard Exportieren")
-        button_header.addWidget(self.neuer_kurs_button)
         button_header.addWidget(self.kurs_löschen_button)
         button_header.addWidget(self.export_scorecard_button) 
         self.layout.addLayout(button_header)
@@ -306,7 +192,7 @@ class HandicapUI(QtWidgets.QWidget):
 
         self.setLayout(self.layout)
 
-        self.neuer_kurs_button.clicked.connect(self.neuer_kurs_hinzugefuegt)
+
         self.kurs_löschen_button.clicked.connect(self.kurs_loeschen_dialog)
         self.export_scorecard_button.clicked.connect(self.oeffne_export_scorecard_dialog) 
 
@@ -326,12 +212,6 @@ class HandicapUI(QtWidgets.QWidget):
 
         self.update()
 
-
-    def neuer_kurs_hinzugefuegt(self):
-        dialog = NeuerKursDialog()
-        if dialog.exec_() == QtWidgets.QDialog.Accepted:
-            dialog.get_kurs_daten()
-            self.update()
 
     def kurs_loeschen_dialog(self):
         kurse = help.getAllCourseIDs()
